refactor(routines): split defaults.rs into module with one file per routine
Converts src/routines/defaults.rs into a defaults/ module.
Each built-in routine gets its own file holding its SPEC const and
prompt text; mod.rs retains DefaultRoutine, DEFAULT_ROUTINES,
materialize, reconcile, and ensure_default_routines.

Closes #702
